Slim down kernel.c
